Created sub-directories for files and moved into, respectively
