+++ src/Web3/Web3Connect/Component/index.js
@@ -1,0 +1,2 @@
+export {default as Web3Button} from './Web3Button';
+export {default as Web3Login} from './Web3Login';
