+++ src/Web3/Web3Connect/Utils/ProviderResource.js
