+++ src/Web3/Web3Connect/Utils/ValidProviders.js
