+++ src/Web3/Web3Connect/Utils/index.js
@@ -1,0 +1,3 @@
+export * from './ProviderResource';
+export * from './useWalletProvider';
+export * from './ValidProviders';
