+++ src/Web3/Web3Connect/index.js
@@ -1,1 +1,2 @@
-export { default as WalletProviderSelection } from "./WalletProviderSelection";
+export * from "./Component";
+export * from "./Utils";
